refactor(settings): split agents section helpers
